moved exceptions to web.errors.py
